node: Create crypto module
